+++ .github/workflows/ingestion-ubwCostumerPerResource.yml
@@ -1,5 +1,5 @@
 
-name: Dataplattform - ingestion - ubwWorkhours
+name: Dataplattform - ingestion - ubwCostumerPerResource
 
 on: 
   push:
@@ -7,7 +7,7 @@ on:
       - master
   pull_request:
     paths:
-      - 'services/ingestion/ubwWorkhours/**'
+      - 'services/ingestion/ubwCostumerPerResource/**'
       - 'packages/common/**'
 
 jobs:
@@ -26,8 +26,8 @@ jobs:
       run: |
         pip install flake8
         # stop the build if there are Python syntax errors or undefined names
-        flake8 ./services/ingestion/ubwWorkhours --count --select=E9,F63,F7,F82 --show-source --statistics
-        flake8 ./services/ingestion/ubwWorkhours --count --max-line-length=120 --statistics
+        flake8 ./services/ingestion/ubwCostumerPerResource --count --select=E9,F63,F7,F82 --show-source --statistics
+        flake8 ./services/ingestion/ubwCostumerPerResource --count --max-line-length=120 --statistics
   
   test:
     runs-on: ubuntu-latest
@@ -37,13 +37,13 @@ jobs:
     steps:
     - uses: actions/checkout@v2
     - name: Install dependencies
-      working-directory: './services/ingestion/ubwWorkhours'
+      working-directory: './services/ingestion/ubwCostumerPerResource'
       run: |
         python -m pip install --upgrade pip
         pip install -r requirements.txt
         pip install -r requirements-test.txt
     - name: Test with pytest
-      working-directory: './services/ingestion/ubwWorkhours'
+      working-directory: './services/ingestion/ubwCostumerPerResource'
       run: |
         pip install pytest
         pytest
