+++ services/ingestion/ubwCostumerPerResource/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "ubwWorkhoursPoller",
+  "name": "ubwCostumerPerResourcePoller",
   "description": "",
   "version": "0.1.0",
   "dependencies": {},
