+++ services/ingestion/ubwCostumerPerResource/pytest.ini
