+++ services/ingestion/ubwCostumerPerResource/requirements-test.txt
