+++ services/ingestion/ubwCostumerPerResource/requirements.txt
