+++ services/ingestion/ubwCostumerPerResource/test_ubw_costumer_per_resource_ingest_lambda.py
@@ -1,4 +1,4 @@
-from ubw_workhours_ingest_lambda import handler
+from ubw_costumer_per_resource_ingest_lambda import handler
 from pytest import fixture
 from json import loads
 
@@ -59,7 +59,7 @@ def zeep_ubw_mock(mocker):
             </Agresso>
         """
     }
-    mocker.patch('ubw_workhours_ingest_lambda.Client', return_value=mock)
+    mocker.patch('ubw_costumer_per_resource_ingest_lambda.Client', return_value=mock)
 
     yield mock
